+++ packages/avatars/src/color/index.ts
@@ -4,6 +4,7 @@ import * as rgbToHex from 'pure-color/convert/rgb2hex';
 import * as hsvToRgb from 'pure-color/convert/hsv2rgb';
 import collection from './collection';
 
+/** @deprecated */
 export interface ColorInterface {
   alpha: number;
   hex: string;
@@ -12,6 +13,7 @@ export interface ColorInterface {
   hsv: number[];
 }
 
+/** @deprecated */
 export default class Color implements ColorInterface {
   public static collection = collection;
 
@@ -30,7 +32,7 @@ export default class Color implements ColorInterface {
       let match = /(.*)\((.*)\)/.exec(color);
 
       if (match) {
-        let values = match[2].split(',').map(val => parseInt(val.trim()));
+        let values = match[2].split(',').map((val) => parseInt(val.trim()));
 
         switch (match[1].trim()) {
           case 'rgb':
@@ -68,7 +70,7 @@ export default class Color implements ColorInterface {
 
     this.alpha = 1;
     this.color = {
-      rgb: rgb
+      rgb: rgb,
     };
   }
 
@@ -96,7 +98,7 @@ export default class Color implements ColorInterface {
 
     this.alpha = 1;
     this.color = {
-      hsv: hsv
+      hsv: hsv,
     };
   }
 
@@ -108,7 +110,7 @@ export default class Color implements ColorInterface {
   set hex(hex: string) {
     this.alpha = 1;
     this.color = {
-      hex: hex
+      hex: hex,
     };
   }
 
@@ -171,14 +173,14 @@ export default class Color implements ColorInterface {
   }
 
   private hexToRgb(hex: string) {
-    return hexToRgb(hex).map(val => Math.round(val));
+    return hexToRgb(hex).map((val) => Math.round(val));
   }
 
   private rgbToHsv(rgb: number[]) {
-    return rgbToHsv(rgb).map(val => Math.round(val));
+    return rgbToHsv(rgb).map((val) => Math.round(val));
   }
 
   private hsvToRgb(hsv: number[]) {
-    return hsvToRgb(hsv).map(val => Math.round(val));
+    return hsvToRgb(hsv).map((val) => Math.round(val));
   }
 }